change webpack to vite, CORS
