--- README.md
@@ -1,29 +1,0 @@
-
-**Project vision. (TeamScript)✨***
-
-### Idea of the project:
-
-Our goal is to create an editor that enables teamwork, project management, and project status monitoring. *Moreover, it allows people to work on the project at the same time.*
-
-Why is this important?
-
-### Realization:
-
-For the convenience of development, it was decided to divide frontend and backend into two different components.
-
-Frontend will be written in (JavaScript), backend in (Java/Python) (maybe something else).
-
-### Features:
-
-Main and most difficult goal for us is to allow users to use our product together. To implement that we will need to use functional DB and to create account
-
-system.
-
-Also, we want to add a few useful features.
-
-- Checking current status of every subproject.
-- Prioritizing tasks will help users to choose how relevant is the specific task at this moment.
-- Creating subtasks’ features will provide users to track big task progress more detailed.
-- *The participants of one project can use our service simultaneously, for example, make changes, add tasks.*
-
-The number of features that can be added to such a project is large. Therefore, as we develop, we will implement more and more features.
