--- website/README.md
@@ -1,24 +1,0 @@
-<h1>Запуск бекенда(backend/)</h1>
-<h4>Установка библиотек</h4>
-
-```
-npm install
-```
-<h4>Запуск сервера</h4>
-
-```
-npm start
-```
-
-<h1>Запуск фронтенда(frontend/)</h1>
-<h4>Установка библиотек</h4>
-
-```
-npm install
-```
-<h4>Запуск сервера</h4>
-
-```
-npm start
-```
-<p>В браузере должна открыться страница http://localhost:3000</p>
