+++ website/backend/package.json
@@ -13,7 +13,10 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "http": "^0.0.1-security",
     "pg": "^8.11.3",
     "socket.io": "^4.7.2",
     "socket.io-client": "^4.7.2"
