--- website/frontend/src/components/TextFieldComponent.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { TextField } from "@mui/material";
-
-
-
-function TextFieldComponent(props) {
-    return (
-          <TextField
-            fullWidth
-            value={props.value}
-            id="standard-basic"
-            label="Enter your text here"
-            multiline
-            onChange={props.updateData}
-          />
-      );
-}
-
-export default TextFieldComponent;
