+++ website/frontend/src/components/TextFieldComponent.jsx
@@ -1,0 +1,16 @@
+import { TextField } from "@mui/material";
+
+function TextFieldComponent(props) {
+  return (
+    <TextField
+      fullWidth
+      value={props.value}
+      id="standard-basic"
+      label="Enter your text here"
+      multiline
+      onChange={props.updateData}
+    />
+  );
+}
+
+export default TextFieldComponent;
